+++ changelog.md
@@ -84,6 +84,7 @@ Codion Change Log
 - AbstractComponentBuilder, Value constructor parameter removed.
 - KeyEvents, overloaded builder() methods removed.
 - FontImageIcon.Builder.IkonBuilder added.
+- TextFieldBuilder.ValueClassBuilder added.
 ### is.codion.swing.framework.model
 - EntityComboBoxModel.Builder.EntityTypeBuilder and ConnectionProviderBuilder added.
 ### is.codion.swing.framework.ui